+++ UI/GameEditorLibrary/Controls/SceneEditing/SceneEditor.cs
@@ -7,6 +7,7 @@
     using Microsoft.Xna.Framework.Graphics;
     using System;
     using System.Windows;
+    using Point = Microsoft.Xna.Framework.Point;
 
     public class SceneEditor : MonoGameViewModel, IGame {
         private readonly EditorCameraWrapper _cameraWrapper;
@@ -18,6 +19,7 @@
         private bool _isInitialized = false;
         private string _rootContentDirectory;
         private IGameSettings _settings;
+        private Point _viewportSize;
 
         public SceneEditor(EditorCameraWrapper cameraWrapper, SelectionEditor selectionEditor) : base() {
             this._cameraWrapper = cameraWrapper;
@@ -27,6 +29,8 @@
         }
 
         public event EventHandler<double> GameSpeedChanged;
+
+        public event EventHandler<Point> ViewportSizeChanged;
 
         public IAssetManager AssetManager {
             get {
@@ -114,6 +118,12 @@
 
         public SpriteBatch SpriteBatch { get; private set; }
 
+        public Point ViewportSize {
+            get {
+                return this._viewportSize;
+            }
+        }
+
         public override void Draw(GameTime gameTime) {
             if (this._isInitialized && this._isContentLoaded) {
                 if (this.CurrentScene != null) {
@@ -177,6 +187,9 @@
         }
 
         public override void SizeChanged(object sender, SizeChangedEventArgs e) {
+            this._viewportSize = new Point(Convert.ToInt32(e.NewSize.Width), Convert.ToInt32(e.NewSize.Height));
+            this.ViewportSizeChanged.SafeInvoke(this, this._viewportSize);
+
             if (e.NewSize.Width > e.PreviousSize.Width || e.NewSize.Height > e.PreviousSize.Height) {
                 this.ResetCamera();
             }